+++ modules/greetd.nix
@@ -1,16 +1,7 @@
 {
   nixosModule = { config, pkgs, lib, ... }: {
-    # Remove this when the programs.regreet.theme options get added
-    # so we no longer need to add these packages manually
-    environment.systemPackages = assert !config.programs.regreet ? theme;
-      builtins.attrValues {
-        inherit (pkgs) gnome-themes-extra;
-        inherit (pkgs.gnome) adwaita-icon-theme;
-      };
-
     services.greetd.enable = true;
     programs.regreet.enable = true;
-    programs.regreet.settings.font_name = "DejaVu Sans 16";
 
     services.greetd.settings.default_session.command =
       "${lib.getExe' pkgs.dbus "dbus-run-session"} ${