+++ modules/linux-builder.nix
@@ -1,8 +1,8 @@
 {
-  darwinModule = { user, host, keys, pkgs, ... }: {
+  darwinModule = { inputs, user, host, keys, pkgs, ... }: {
     nix.linux-builder.enable = true;
     nix.linux-builder.package = pkgs.darwin.linux-builder;
-    nix.linux-builder.config = { config, pkgs, ... }: {
+    nix.linux-builder.config = { config, pkgs, lib, ... }: {
       networking.hostName = "${host}-linux-builder";
 
       services.tailscale.enable = true;
@@ -27,8 +27,10 @@
 
       nix.settings.experimental-features = "nix-command flakes";
       nix.settings.secret-key-files = [ "/etc/nix/key" ];
-    };
 
-    nix.settings.extra-trusted-users = [ user ];
+      nix.settings.trusted-users = lib.mkForce [ "root" ];
+
+      nixpkgs.overlays = [ inputs.nix-overlay.overlay ];
+    };
   };
 }